requirements.txt
reportlab==4.0.4
openai-whisper==20231117
torch>=2.0.0 